+++ Sources/Runner/main.swift
@@ -3,7 +3,7 @@ import Danger
 
 let cliLength = ProcessInfo.processInfo.arguments.count
 do {
-    let isVerbose = CommandLine.arguments.contains("--verbose")
+    let isVerbose = CommandLine.arguments.contains("--verbose") || ProcessInfo.processInfo.environment["DEBUG"] != nil
     let isSilent = CommandLine.arguments.contains("--silent")
     let logger = Logger(isVerbose: isVerbose, isSilent: isSilent)
     if cliLength > 1 && "edit" == CommandLine.arguments[1] {